requirements.txt
asgiref==3.7.2
azure-core==1.29.3
azure-servicebus==7.11.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
dnspython==2.4.2
gunicorn==21.2.0
idna==3.4
isodate==0.6.1
packaging==23.1
psycopg2-binary==2.9.9
pymongo==4.5.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
